Move Combine and Promise test extensions to test target
